+++ Grana/Source/TrapezoidalEnvelope.h
@@ -24,20 +24,20 @@ public:
     
 
     static void reset();
-    static void reset(float duration, int sampleRate, float mainLobeWidth);
+    static void reset(int duration, int sampleRate, float mainLobeWidth);
 
     static TrapezoidalEnvelope* getInstance();
     static TrapezoidalEnvelope* setSampleRate(int sampleRate);
     static TrapezoidalEnvelope* setMainLobeWidth(int mainLobeWidth);
     static TrapezoidalEnvelope* setDuration(int duration);
-
+    
     
 
 private:
     TrapezoidalEnvelope();
     TrapezoidalEnvelope(int sampleRate);
-    TrapezoidalEnvelope(float duration, int sampleRate);
-    TrapezoidalEnvelope(float duration, int sampleRate, float mainLobeWidth);
+    TrapezoidalEnvelope(int duration, int sampleRate);
+    TrapezoidalEnvelope(int duration, int sampleRate, float mainLobeWidth);
     ~TrapezoidalEnvelope();
 
     void filterCreation();
